[Mod] 64bitに対応。
git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/xspfmanager/XspfManager@259 ae6f0248-dcd5-41b5-8401-a3f1d2e3b736

+++ HMChannel.h
@@ -20,7 +20,7 @@
 	NSInteger _requestNum;
 }
 
--(id)initWithWorkerNum:(int)num;
+-(id)initWithWorkerNum:(NSInteger)num;
 
 // called from HMWorker. take care thread safty.
 - (oneway void)finishRequest:(id <HMRequest>)request;
